+++ data/romanian-tree.js
@@ -13,7 +13,7 @@ import { NounTemplateNeuter,
     ReflexiveVerbTemplateSi,
     ReflexiveVerbTemplateSe,
     VerbTemplate
-} from "../src/romanianLanguageModels.js";
+} from "./romanian-language-models.js";
 
 /*
 a ă â b c d e f g h i î j k l m n o p q r s ș t ț u v w x y z
